+++ lib/router.js
@@ -1576,7 +1576,7 @@ router.get('/asahi/area/:id', lazyloadRouteHandler('./routes/asahi/area'));
 router.get('/asahi/:genre?/:category?', lazyloadRouteHandler('./routes/asahi/index'));
 
 // 7x24小时快讯
-router.get('/fx678/kx', lazyloadRouteHandler('./routes/fx678/kx'));
+// router.get('/fx678/kx', lazyloadRouteHandler('./routes/fx678/kx'));
 
 // SoundCloud
 router.get('/soundcloud/tracks/:user', lazyloadRouteHandler('./routes/soundcloud/tracks'));
