--- lib/routes/fx678/kx.js
@@ -1,51 +1,0 @@
-const cheerio = require('cheerio');
-const got = require('@/utils/got');
-const url = require('url');
-const { parseDate } = require('@/utils/parse-date');
-const timezone = require('@/utils/timezone');
-const host = 'https://www.fx678.com/kx/';
-
-module.exports = async (ctx) => {
-    const link = 'https://www.fx678.com/kx/';
-    const res = await got.get(link);
-    const $ = cheerio.load(res.data);
-    // 页面新闻消息列表
-    const list = $('.body_zb ul .body_zb_li .zb_word')
-        .find('.list_font_pic > a:first-child')
-        .map((i, e) => $(e).attr('href'))
-        .slice(0, 30)
-        .get();
-
-    const out = await Promise.all(
-        list.map(async (itemUrl) => {
-            const absoluteUrl = url.resolve(host, itemUrl);
-            const cache = await ctx.cache.get(absoluteUrl);
-            // 判断缓存
-            if (cache) {
-                return Promise.resolve(JSON.parse(cache));
-            }
-            const res = await got.get(absoluteUrl);
-            const $ = cheerio.load(res.data);
-
-            const contentPart = $('.article-main .content').html().trim();
-            const forewordPart = $('.article-main .foreword').html().trim();
-            const datetimeString = $('.article-cont .details i').text().trim();
-            const articlePubDate = timezone(parseDate(datetimeString, 'YYYY-MM-DD HH:mm:ss'), +8);
-
-            const item = {
-                title: $('.article-main .foreword').text().trim().split('——').pop(),
-                link: absoluteUrl,
-                description: contentPart.length > 1 ? contentPart : forewordPart,
-                pubDate: articlePubDate,
-            };
-            ctx.cache.set(absoluteUrl, JSON.stringify(item));
-
-            return item;
-        })
-    );
-    ctx.state.data = {
-        title: '7x24小时快讯',
-        link,
-        item: out,
-    };
-};
