+++ lib/v2/fx678/maintainer.js
@@ -1,0 +1,3 @@
+module.exports = {
+    '/kx': ['occupy5', 'dousha'],
+};
