+++ lib/v2/fx678/radar.js
@@ -1,0 +1,13 @@
+module.exports = {
+    'fx678.com': {
+        _name: '汇通网',
+        '.': [
+            {
+                title: '7x24 小时快讯',
+                docs: 'https://docs.rsshub.app/finance.html#hui-tong-wang',
+                source: ['/kx'],
+                target: '/fx678/kx',
+            },
+        ],
+    },
+};
