+++ lib/v2/fx678/router.js
@@ -1,0 +1,3 @@
+module.exports = (router) => {
+    router.get('/kx', require('./kx'));
+};
